requirements.txt
numpy
# scipy
pillow
numba